+++ offline/framework/ffarawobjects/MvtxRawEvtHeader.h
@@ -27,14 +27,14 @@ public:
   void AddFeeId(const int& feeid) { m_MvtxFeeIdSet.insert(feeid); };
   void AddL1Trg(const uint64_t& gtmL1_bco) { m_MvtxL1TrgSet.insert(gtmL1_bco); };
 
-  void AddFeeId(const std::set<int>& mvtxFeeIds);
+  void AddFeeId(const std::set<uint16_t>& mvtxFeeIds);
   void AddL1Trg(const std::set<uint64_t>& mvtxL1TrgSet);
 
-  std::set<int>& getMvtxFeeIdSet() { return m_MvtxFeeIdSet; };
+  std::set<uint16_t>& getMvtxFeeIdSet() { return m_MvtxFeeIdSet; };
   std::set<uint64_t>& getMvtxLvL1BCO() { return m_MvtxL1TrgSet; };
 
 private:
-  std::set<int> m_MvtxFeeIdSet;
+  std::set<uint16_t> m_MvtxFeeIdSet;
   std::set<uint64_t> m_MvtxL1TrgSet;
 
   ClassDefOverride(MvtxRawEvtHeader, 2)
